chore: create socket file in cwd
