+++ packages/playwright/src/mcp/terminal/cli.ts
@@ -222,9 +222,25 @@ class SocketSession {
 }
 
 function daemonSocketPath(): string {
-  if (os.platform() === 'win32')
-    return path.join('\\\\.\\pipe', 'pw-daemon.sock');
-  return path.join(os.homedir(), '.playwright', 'pw-daemon.sock');
+  const socketPath = path.resolve('.playwright.sock');
+  return normalizeSocketPath(socketPath);
+}
+
+/**
+ * Normalize socket path for the current platform.
+ * On Windows, converts Unix-style paths to named pipe format.
+ * On Unix, returns the path as-is.
+ */
+function normalizeSocketPath(path: string): string {
+  if (os.platform() === 'win32') {
+    // Windows named pipes use \\.\pipe\name format
+    if (path.startsWith('\\\\.\\pipe\\'))
+      return path;
+    // Convert Unix-style path to Windows named pipe
+    const name = path.replace(/[^a-zA-Z0-9]/g, '-');
+    return `\\\\.\\pipe\\${name}`;
+  }
+  return path;
 }
 
 async function connectToDaemon(): Promise<SocketSession> {
@@ -246,6 +262,7 @@ async function connectToDaemon(): Promise<SocketSession> {
   const child = spawn(process.execPath, [cliPath, 'run-mcp-server', `--daemon=${socketPath}`], {
     detached: true,
     stdio: 'ignore',
+    cwd: process.cwd(), // Will be used as root.
   });
   child.unref();
